Move the hython launcher to zabob.common
