+++ zabob-modules/src/zabob/core/main.py
@@ -10,11 +10,11 @@ import webbrowser
 import click
 import os
 
-from zabob.core.diagnostics import diagnostics
-from zabob.core.paths import ZABOB＿MARKSERV, ZABOB_ROOT
-from zabob.core.hython import hython
 from zabob.common.common_utils import INFO, QUIET, SILENT, VERBOSE, Level, DEBUG
 from zabob.common.subproc import run
+from zabob.common.hython import hython
+from zabob.core.diagnostics import diagnostics
+from zabob.core.paths import ZABOB＿MARKSERV, ZABOB_ROOT
 
 
 @click.group()
